feat(mzpeak): chunk chromatograms over the time axis (reference parity)
chromatograms_data now follows the spectra layout in the chunked modes: one
time-axis chunk per chromatogram mirroring the HUPO-PSI reference struct
(chromatogram_index, time_chunk_start/end, time_chunk_values, chunk_encoding,
intensity, [time_numpress_linear_bytes]). Default numpress-linear time
(MS:1002312); --lossless delta (MS:1003089); intensity widened to f64 to match
the reference. --point keeps the legacy per-point chromatogram (with ms_level).

New ChromChunkFacetStream + IChromDataFacet; CaptureTic now carries f64 (the
point facet still narrows to f32, so --point output is byte-identical). The
chunk chromatogram carries no array_index (the reference omits it; the validator
infers layout from column names). Adds 2 tests (numpress shape; lossless delta
time round-trips exactly vs the point layout). 97/97 native arm64; 3-mode
mzpeak-validate 0 errors. Adds Phase 5 CERT.

+++ ThermoRawFileParser/Writer/MzPeak/ScanStager.cs
@@ -165,7 +165,7 @@ namespace ThermoRawFileParser.Writer
         }
 
         private void CaptureTic(IRawDataPlus raw, List<MzPeakRecord> records, List<double> time,
-            List<float> intensity, List<long> msLevel)
+            List<double> intensity, List<long> msLevel)
         {
             ChromatogramSignal[] trace;
             if (TestForceSummedTic)
@@ -206,7 +206,7 @@ namespace ThermoRawFileParser.Writer
                 for (int i = 0; i < records.Count; i++)
                 {
                     time.Add(signal.Times[i]);
-                    intensity.Add((float)signal.Intensities[i]);
+                    intensity.Add(signal.Intensities[i]);
                     msLevel.Add(records[i].MsLevel);
                 }
                 _chromFromDeviceTrace = true;
@@ -225,7 +225,7 @@ namespace ThermoRawFileParser.Writer
                 if (byScan.TryGetValue(r.ScanNumber, out var sample))
                 {
                     time.Add(sample.time);
-                    intensity.Add((float)sample.intensity);
+                    intensity.Add(sample.intensity);
                 }
                 else
                 {
